Start multiprocessing with another script
