+++ .gitignore
@@ -14,3 +14,5 @@
 
 # TensorBoard log
 /tensorboard/**/*
+
+__pycache__
